+++ packages/vm/src/cache.rs
@@ -24,6 +24,7 @@ struct Stats {
 pub struct CosmCache<S: Storage + 'static, A: Api + 'static, Q: Querier + 'static> {
     wasm_path: PathBuf,
     supported_features: HashSet<String>,
+    print_debug: bool,
     modules: FileSystemCache,
     stats: Stats,
     // Those two don't store data but only fix type information
@@ -50,6 +51,7 @@ where
     pub unsafe fn new<P: Into<PathBuf>>(
         base_dir: P,
         supported_features: HashSet<String>,
+        print_debug: bool,
     ) -> VmResult<Self> {
         let base = base_dir.into();
         let wasm_path = base.join(WASM_DIR);
@@ -61,6 +63,7 @@ where
         Ok(CosmCache {
             wasm_path,
             supported_features,
+            print_debug,
             modules,
             stats: Stats::default(),
             type_storage: PhantomData::<S>,
@@ -104,13 +107,13 @@ where
         let res = self.modules.load_with_backend(checksum, backend());
         if let Ok(module) = res {
             self.stats.hits_module += 1;
-            return Instance::from_module(&module, deps, gas_limit);
+            return Instance::from_module(&module, deps, gas_limit, self.print_debug);
         }
 
         // fall back to wasm cache (and re-compiling) - this is for backends that don't support serialization
         let wasm = self.load_wasm(checksum)?;
         self.stats.misses += 1;
-        Instance::from_code(&wasm, deps, gas_limit)
+        Instance::from_code(&wasm, deps, gas_limit, self.print_debug)
     }
 }
 
@@ -173,7 +176,7 @@ mod test {
     fn save_wasm_works() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         cache.save_wasm(CONTRACT).unwrap();
     }
 
@@ -182,7 +185,7 @@ mod test {
     fn save_wasm_allows_saving_multiple_times() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         cache.save_wasm(CONTRACT).unwrap();
         cache.save_wasm(CONTRACT).unwrap();
     }
@@ -203,7 +206,7 @@ mod test {
 
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let save_result = cache.save_wasm(&wasm);
         match save_result.unwrap_err() {
             VmError::StaticValidationErr { msg, .. } => {
@@ -217,7 +220,7 @@ mod test {
     fn load_wasm_works() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         let restored = cache.load_wasm(&id).unwrap();
@@ -232,13 +235,13 @@ mod test {
 
         {
             let mut cache1: CosmCache<MockStorage, MockApi, MockQuerier> =
-                unsafe { CosmCache::new(tmp_path, default_features()).unwrap() };
+                unsafe { CosmCache::new(tmp_path, default_features(), false).unwrap() };
             id = cache1.save_wasm(CONTRACT).unwrap();
         }
 
         {
             let cache2: CosmCache<MockStorage, MockApi, MockQuerier> =
-                unsafe { CosmCache::new(tmp_path, default_features()).unwrap() };
+                unsafe { CosmCache::new(tmp_path, default_features(), false).unwrap() };
             let restored = cache2.load_wasm(&id).unwrap();
             assert_eq!(restored, CONTRACT);
         }
@@ -248,7 +251,7 @@ mod test {
     fn load_wasm_errors_for_non_existent_id() {
         let tmp_dir = TempDir::new().unwrap();
         let cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let checksum = Checksum::from([
             5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5,
             5, 5, 5,
@@ -267,7 +270,7 @@ mod test {
     fn load_wasm_errors_for_corrupted_wasm() {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache: CosmCache<MockStorage, MockApi, MockQuerier> =
-            unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let checksum = cache.save_wasm(CONTRACT).unwrap();
 
         // Corrupt cache file
@@ -286,7 +289,8 @@ mod test {
     #[test]
     fn get_instance_finds_cached_module() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         let deps = mock_dependencies(20, &[]);
         let _instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
@@ -297,7 +301,8 @@ mod test {
     #[test]
     fn get_instance_finds_cached_instance() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         let deps1 = mock_dependencies(20, &[]);
         let deps2 = mock_dependencies(20, &[]);
@@ -312,7 +317,8 @@ mod test {
     #[test]
     fn init_cached_contract() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         let deps = mock_dependencies(20, &[]);
         let mut instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
@@ -330,7 +336,8 @@ mod test {
     #[test]
     fn run_cached_contract() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
         // TODO: contract balance
         let deps = mock_dependencies(20, &[]);
@@ -354,7 +361,8 @@ mod test {
     #[test]
     fn use_multiple_cached_instances_of_same_contract() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         // these differentiate the two instances of the same contract
@@ -400,7 +408,8 @@ mod test {
     #[cfg(feature = "default-singlepass")]
     fn resets_gas_when_reusing_instance() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         let deps1 = mock_dependencies(20, &[]);
@@ -431,7 +440,8 @@ mod test {
     #[cfg(feature = "default-singlepass")]
     fn recovers_from_out_of_gas() {
         let tmp_dir = TempDir::new().unwrap();
-        let mut cache = unsafe { CosmCache::new(tmp_dir.path(), default_features()).unwrap() };
+        let mut cache =
+            unsafe { CosmCache::new(tmp_dir.path(), default_features(), false).unwrap() };
         let id = cache.save_wasm(CONTRACT).unwrap();
 
         let deps1 = mock_dependencies(20, &[]);